move low level op.Box into `op.py` for consistency
