+++ tests/models/test_box_map.py
@@ -4,8 +4,8 @@ from collections.abc import Generator
 import algopy
 import pytest
 from algopy_testing import arc4
-from algopy_testing.box import BoxMap
 from algopy_testing.context import AlgopyTestContext, algopy_testing_context
+from algopy_testing.models.box import BoxMap
 from algopy_testing.primitives.biguint import BigUInt
 from algopy_testing.primitives.bytes import Bytes
 from algopy_testing.primitives.string import String
